Update consensus_calling_without_umis
The first part -split file- splits the big file into 40 segments each time, to do parallel processing, in each segment based on the length of the sequence picked it decides a mismatch and filters out the reads for the ref seq that pass the threshold of the mismatch and make file for that sequences with sequences passing the threshold. 

Then in second part each of this file is used to call a consensus sequence for them.